requirements.txt
# ── Core ML ───────────────────────────────────────────────────────────────────
sentence-transformers==3.3.1
transformers==4.47.1
torch==2.5.1
Pillow==11.0.0

# ── Vision / CLIP ─────────────────────────────────────────────────────────────
# Uses HuggingFace transformers CLIPModel + CLIPProcessor (no extra package needed)

# ── Vector Store ──────────────────────────────────────────────────────────────
chromadb==0.6.3

# ── Sparse Retrieval ──────────────────────────────────────────────────────────
rank-bm25==0.2.2

# ── Video Processing ──────────────────────────────────────────────────────────
opencv-python-headless==4.10.0.84

# ── LLM Client (Ollama HTTP API) ──────────────────────────────────────────────
httpx==0.28.1

# ── Observability ─────────────────────────────────────────────────────────────
prometheus-client==0.21.1
opentelemetry-sdk==1.29.0
opentelemetry-api==1.29.0

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.41.1

# ── CLI ───────────────────────────────────────────────────────────────────────
tqdm==4.67.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
anyio==4.7.0

# ── Dev utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
numpy==1.26.4
